feat(products): 모델의 status 이름 변경, 내 등록상품 조회 코드 수정

+++ products/serializers.py
@@ -35,17 +35,24 @@ class ProductSerializer(serializers.ModelSerializer):
         return product
     
     def update(self, instance, validated_data):
-        uploaded_images = validated_data.pop('uploaded_images')
         change_name = validated_data.get('product_name', None)
         
         if change_name:
             instance.product_name = change_name
-        # 이전 이미지 삭제
-        instance.images.all().delete()
-        # 새로 이미지 등록
-        for image in uploaded_images:
-            ProductImage.objects.create(product=instance, image=image)
+        
+        if 'uploaded_images' in validated_data:
+            uploaded_images = validated_data.pop('uploaded_images')
+            # 이전 이미지 삭제
+            instance.images.all().delete()
+            # 새로 이미지 등록
+            for image in uploaded_images:
+                ProductImage.objects.create(product=instance, image=image)
 
         return super().update(instance, validated_data)
+
+
+
+
+
